Extract generic request/response functions to Api module
